Ran cygwin firstsetup and fixed bugs
Don't create home link if unnecessary.

Don't link usrlocalbin -- put it in PATH instead.

+++ win/cygwin/firstsetup/cygwin_links.sh
@@ -6,9 +6,15 @@ cd /
 
 ln -s --no-target-directory /cygdrive/c /c
 
-mv /home/$USER /home/$USER_old
-ln -s --no-target-directory /cygdrive/c/Users/$USER /home/$USER
+if [ ! -e /home/$USER/data ] ;
+	# I expect daveconfig's root links have been created. If the cygwin home
+	# directory doesn't contain our data directory, then it's not pointing at
+	# the Windows home. Fix that.
+	mv /home/$USER /home/$USER_old
+	ln -s --no-target-directory /cygdrive/c/Users/$USER /home/$USER
+fi
 # May need to manually move ~/davidb into /home/davidb
 
-mv /usr/local/bin /usrlocalbin_old
-ln -s --no-target-directory /cygdrive/c/david/settings/daveconfig/win/cygwin/usrlocalbin /usr/local/bin
+# Instead of this, add usrlocalbin to path.
+# mv /usr/local/bin /usrlocalbin_old
+# ln -s --no-target-directory /cygdrive/c/david/settings/daveconfig/win/cygwin/usrlocalbin /usr/local/bin
